Move unification functions into separate file
